auth/requirements.txt
aioredis==2.0.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncio==4.0.0
asyncpg==0.30.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
fastapi==0.120.4
h11==0.16.0
idna==3.11
loguru==0.7.3
passlib==1.7.4
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.0.1
sniffio==1.3.1
starlette==0.49.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
